Split partitioning logic into multiple modules
The items are reexported to retain the apparent structure.
